% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_query.R
\name{define_query}
\alias{define_query}
\title{Define ERIC query}
\usage{
define_query(args)
}
\arguments{
\item{args}{takes args, normally passed from search_eric}
}
\value{
q_string
}
\description{
Define ERIC query
}
\examples{
define_query(list(search = "epistemic", author = "Knight"))
}
