% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_eric.R
\name{search_eric}
\alias{search_eric}
\title{Main function to search eric and return results}
\usage{
search_eric(
  search = NULL,
  title = NULL,
  author = NULL,
  source = NULL,
  subject = NULL,
  description = NULL,
  start = 0,
  rows = 100,
  check = FALSE,
  fields = c("author", "description", "id", "isbn", "issn", "language", "peerreviewed",
    "publicationdateyear", "publicationtype", "publisher", "subject", "title", "url"),
  base_url = "https://api.ies.ed.gov/eric/"
)
}
\arguments{
\item{search}{general search, can use fields
(title, author, source, subject, or description).
Function assumes that AND is used to combine search fields where multiple are specified;
If AND is not the intended operator, manually include the fields in the 'search' variable}

\item{title}{search-field}

\item{author}{search-field}

\item{source}{search-field}

\item{subject}{search-field}

\item{description}{search-field}

\item{start}{the record number to start from. Default 0}

\item{rows}{the number of records to return, default 100, max 1000, set to 0 for just meta}

\item{check}{whether to check before returning results}

\item{fields}{the fields to return.
If set to NULL it returns c(author, description, id, isbn, issn, language, peerreviewed, publicationdateyear, publicationtype, publisher, subject, title).
Full list is c(abstractor, audience, authorxlink, e_datemodified, e_fulltextauth, e_yearadded, educationlevel, identifiersgeo, identifierslaw, identifierstest, iescited, iesfunded, iesgrantconstractnum, iesgrantcontractnumxlink, ieslinkdatasource, ieslinkpublication, ieslinkwwcreviewguide, ieswwcreviewed, institution, sponsor, sourceid, source, url)}

\item{base_url}{unlikely to change, default is valid.}
}
\value{
content_new
}
\description{
Simple wrapper for the ERIC search API, main function, returns a dataframe and
some metadata to console. Documentation is largely restating the ERIC API materials.
Education Resources Information Center,
"ERIC is a comprehensive, easy-to-use, searchable,
Internet-based bibliographic and full-text database of education research and information"
}
\examples{
search_eric(search = "epistemic cognition", author = "Knight, S")
}
